fix: separate wallet / connections
