+++ frontend/src/components/AppHeader.tsx
@@ -6,10 +6,10 @@ type Props = {
 
 function AppHeader({ title, description, contentRight }: Props) {
   return (
-    <div className="flex justify-between border-b border-border pb-6">
+    <div className="flex justify-between border-b border-border pb-3 lg:pb-6">
       <div className="flex-1">
         <h1 className="text-xl lg:text-3xl font-semibold">{title}</h1>
-        <p className="text-muted-foreground">{description}</p>
+        <p className="hidden lg:inline text-muted-foreground">{description}</p>
       </div>
       <div className="flex gap-3">{contentRight}</div>
     </div>
