+++ frontend/src/components/redirects/HomeRedirect.tsx
@@ -17,7 +17,7 @@ export function HomeRedirect() {
     if (info.setupCompleted && info.running) {
       if (info.unlocked) {
         const returnTo = window.localStorage.getItem(localStorageKeys.returnTo);
-        to = returnTo || "/connections";
+        to = returnTo || "/wallet";
       } else {
         to = "/unlock";
       }
